argus-incident-responder/backend/requirements.txt
# Generated by uv — do not edit manually.
# To regenerate: uv pip compile pyproject.toml -o requirements.txt
# To install:    uv pip sync requirements.txt
#
# Source of truth is pyproject.toml.
# This file is kept for Docker builds and CI environments that use pip.

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
google-genai>=0.5.0
google-cloud-bigquery>=3.20.0
google-cloud-firestore>=2.16.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
faker>=25.0.0
